Change TransactionDialog for TransactionCardDetail.tsx
