+++ app/src/components/NavBar.tsx
@@ -10,7 +10,7 @@ export default function NavBar() {
         <TurtleIcon size={40} />
         <div className="turtle-text-shadow xl-letter-spacing text-[2rem] text-white">Turtle</div>
         <div className="box-shadow mt-[1px] flex h-[19px] items-center justify-center rounded-[4px] border border-white bg-white px-[5px]">
-          <Tooltip showIcon={false} content={'Turtle now supports XCM transfers 🚀🫶🏻🥳'}>
+          <Tooltip showIcon={false} content="Turtle now supports XCM transfers 🚀🫶🏻🥳">
             <p className="rainbow-text text-[12px] font-bold">v2.0.0</p>
           </Tooltip>
         </div>
